# framezzz-api/cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api Endpoints for Go Server",
        "title": "Go Rest Api",
        "contact": {}
    },
    "paths": {
...
        "/api/comic/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comics"
                ],
                "summary": "Get a specific comic",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetComicResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.GetComicResponse": {
            "type": "object",
            "properties": {
                "coverUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "postUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}